convert the flow from data to IO thread

+++ app/src/main/java/com/example/marvel_application/model/remote/response/comic/Price.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Price(
     @SerializedName("price")
-    val price: Int? = null,
+    val price: Double? = null,
     @SerializedName("type")
     val type: String? = null
 )
